requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
sqlalchemy==2.0.35
psycopg[binary]>=3.2.1
stripe==11.1.1
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.12
httpx==0.27.2

# Security dependencies
PyJWT==2.8.0
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==10.4.0

# Testing dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==4.1.0
